implementation of knapsack algorithm additional  endpoints: loadDroneWithMedication ,loadDroneWithMedication,checkDroneBatteryLevel endpoints in DroneController Medication business logic added sample test data and endpoints in postman/insomnia collection files. added more sql scripts in data.sql file

+++ src/main/java/com/musala/drone/task/dto/DroneBatteryLevelResponse.java
@@ -9,6 +9,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class DroneBatteryLevelResponse {
 	
-	public int batteryCapacityLevel ;
+	public Double batteryCapacityLevel ;
 
 }
